requirements.txt
accelerate
arabic-reshaper
beautifulsoup4
craft-text-detector
datasets
evaluate
fastapi
gdown
google-generativeai
gradio
huggingface-hub
jiwer
kornia
matplotlib
numpy==1.26.4
opencv-python
pandas
peft
pillow
python-bidi
python-dotenv
qwen-vl-utils
rapidfuzz
requests
rich
safetensors
scipy
transformers==4.47.1
tqdm
typer
uvicorn